+++ examples/resources/morpheus_task_job/resource_schedule.tf
@@ -9,6 +9,7 @@ data "morpheus_execute_schedule" "example_schedule" {
 resource "morpheus_task_job" "tf_example_task_job_schedule" {
   name                  = "TF Example Task Job Schedule"
   enabled               = true
+  labels                = ["aws", "demo"]
   task_id               = data.morpheus_task.example_task.id
   schedule_mode         = "scheduled"
   execution_schedule_id = data.morpheus_execute_schedule.jobtest.id